+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 - Fixed recall platforms working without electricity
 - Fixed shade sword altar item icon facing the wrong direction
 - Fixed mankh and cloak spawning not consuming their respective item
+- Fixed raid crashes caused by the illager golem
 - Sacred Pontiff is now respawnable by interacting on a respawn totem with a nether star
 - Increased Sacred Pontiff randomized loot amount
+- Now the illager golem spawns once per even raid wave
 
 # 1.1.2
 - Added compat with knightlib 1.4.0